+++ GitTfs/Core/TfsChangesetInfo.cs
@@ -6,7 +6,7 @@ namespace Sep.Git.Tfs.Core
     public class TfsChangesetInfo
     {
         public IGitTfsRemote Remote { get; set; }
-        public long ChangesetId { get; set; }
+        public int ChangesetId { get; set; }
         public string GitCommit { get; set; }
         public IEnumerable<ITfsWorkitem> Workitems { get; set; }
